test(check): the validation reports become spec documents
Thirty tests over five files. Twenty-three are documents now, and the
gain is not brevity: `spwn check` is tool output, so its report belongs in
a full golden per scoped use case (rule D11) rather than in the grep
clusters five of these specs had grown. The invalid-tool report now pins
the built-in catalogue it offers; the legacy-refs report pins all three
refusals and the grammar each one prints.

Three specs stay in code and say why in their own file. `check --json` is
judged STRUCTURALLY — `expect(result.json).toMatch('valid.json')` compares
by shape, and the compile pass is filtered by a field; `stdout:` is
byte-exact prose and has no vocabulary for either.

Two more use the bridge: their subject is what is ABSENT from agent.yaml
(no entry may survive bare), and `files:` says what a file contains, never
what it does not. The session stays in its document, `cli.run()` asserts
it whole, and the code adds the one probe.

Dropped as subsumed: four validation specs that re-probed by substring a
command another document now asserts byte-exact — the two usage-dump
absence probes, the actionable-hint probe, and a second `unknown command`
case, plus the github-registry probe that duplicated check's own golden.
