+++ tests/specs/cli/check/absent-lockfile.spec.yaml
@@ -1,0 +1,10 @@
+description: the lockfile rule stays silent when no lockfile exists
+fixture: $FIXTURES/docker-pilot/
+runs:
+    - command: check
+      exit: 0
+      stdout: |4+
+          
+            ✓  Project is valid
+               {{workdir}}/spwn.yaml
+
