+++ tests/specs/cli/check/complete-lockfile.spec.yaml
@@ -1,0 +1,20 @@
+description: check passes once every declared ref is installed
+fixture: $FIXTURES/docker-pilot/
+runs:
+    - command: install spwn:unix
+      exit: 0
+      stdout: |4
+            ✓  installed spwn:unix
+               1 agent updated, spwn.lock pinned
+    - command: install spwn:git
+      exit: 0
+      stdout: |4
+            ✓  installed spwn:git
+               1 agent updated, spwn.lock pinned
+    - command: check
+      exit: 0
+      stdout: |4+
+          
+            ✓  Project is valid
+               {{workdir}}/spwn.yaml
+
