+++ tests/specs/cli/check/drifting-lockfile.spec.yaml
@@ -1,0 +1,21 @@
+description: an incomplete lockfile is flagged by name
+fixture: $FIXTURES/docker-pilot/
+runs:
+    - command: install spwn:unix
+      exit: 0
+      stdout: |4
+            ✓  installed spwn:unix
+               1 agent updated, spwn.lock pinned
+    - command: check
+      exit: 1
+      stdout: |4+
+          
+            {{workdir}}/spwn.yaml
+
+            errors
+              •  "spwn:git" is not recorded in spwn.lock
+                 spwn.yaml#deps
+                 → run `spwn install spwn:git` to sync the lockfile
+
+            1 error(s), 0 warning(s), 0 info
+
