+++ tests/specs/cli/check/duplicate-world-claim.spec.yaml
@@ -1,0 +1,25 @@
+description: a second world claiming the same agent breaks the one-agent-one-world rule
+fixture:
+    - $FIXTURES/single-agent/
+    - two-worlds-same-agent/
+runs:
+    - command: check
+      exit: 1
+      stdout: |4+
+          
+            {{workdir}}/spwn.yaml
+
+            errors
+              •  agent "neo" already deployed by world "duplicate"
+                 spwn.yaml#worlds.neo.agents
+                 → an agent may only belong to a single world; remove the duplicate
+
+            info
+              •  world "duplicate" has no knowledge path; agents will see an empty /world/knowledge/ and won't be told one exists.
+                 spwn.yaml#worlds.duplicate.knowledge
+                 → add `knowledge: ./knowledge` (or another path) to enable the shared knowledge base for this world
+
+            1 error(s), 0 warning(s), 1 info
+
+      files:
+          spwn.yaml: { contains: "duplicate:" }
