+++ tests/specs/cli/check/empty-agent-prompt.spec.yaml
@@ -1,0 +1,25 @@
+description: an empty AGENTS.md passes the shallow check and fails the deep one
+fixture:
+    - $FIXTURES/single-agent/
+    - empty-agents-md/
+runs:
+    - command: check
+      exit: 0
+      stdout: |4+
+          
+            ✓  Project is valid
+               {{workdir}}/spwn.yaml
+
+    - command: check --deep
+      exit: 1
+      stdout: |4+
+          
+            {{workdir}}/spwn.yaml
+
+            compile
+              •  agent prompt is missing or empty
+                 spwn/agents/neo/AGENTS.md
+                 → create spwn/agents/neo/AGENTS.md with the agent's system prompt
+
+            1 error(s), 0 warning(s), 0 info
+
