--- tests/specs/cli/check/expected/help.txt
@@ -1,15 +1,0 @@
-
-Walks up from the current directory looking for spwn.yaml, then runs
-every validation rule against the project. Reports issues grouped by
-severity. Exits non-zero when errors are found (or warnings, with
---strict).
-
-Usage:
-  spwn check [flags]
-
-Flags:
-      --deep     Also run a compile dry-run and report renderer errors
-  -h, --help     help for check
-      --json     Emit results as structured JSON on stdout
-      --strict   Exit non-zero on warnings, not just errors
-
