--- tests/specs/cli/check/expected/invalid-tool-ref.txt
@@ -1,10 +1,0 @@
-
-  {{workdir}}/spwn.yaml
-
-  errors
-    •  dependency "spwn:nonexistent" does not exist
-       spwn/agents/neo/agent.yaml#deps
-       → available built-ins: spwn:architect, spwn:build, spwn:cli, spwn:docker-cli, spwn:gh, spwn:git, spwn:mcp2cli, spwn:mempalace, spwn:node, spwn:notion, spwn:python, spwn:qmd, spwn:unix, spwn:x, spwn:claude-code, spwn:codex
-
-  1 error(s), 0 warning(s), 0 info
-
