--- tests/specs/cli/check/expected/valid-project.txt
@@ -1,4 +1,0 @@
-
-  ✓  Project is valid
-     {{workdir}}/spwn.yaml
-
