+++ tests/specs/cli/check/legacy-dependency-refs.spec.yaml
@@ -1,0 +1,22 @@
+description: bare, local and @owner dependency refs are each rejected with the grammar
+fixture: $FIXTURES/check-legacy-refs/
+runs:
+    - command: check
+      exit: 1
+      stdout: |4+
+          
+            {{workdir}}/spwn.yaml
+
+            errors
+              •  dependency "python" is invalid
+                 spwn/agents/neo/agent.yaml#deps
+                 → use skill/<name> (for spwn/skills/<name>.md), tool/<name> (for spwn/tools/<name>/), hook/<name> (for spwn/hooks/<name>.yaml), or command/<name> (for spwn/commands/<name>.md); spwn:<name> and github:<owner>/<repo> remain the two source-prefixed schemes
+              •  dependency "local:foo" is invalid
+                 spwn/agents/neo/agent.yaml#deps
+                 → use skill/<name> (for spwn/skills/<name>.md), tool/<name> (for spwn/tools/<name>/), hook/<name> (for spwn/hooks/<name>.yaml), or command/<name> (for spwn/commands/<name>.md); spwn:<name> and github:<owner>/<repo> remain the two source-prefixed schemes
+              •  dependency "@acme/foo" is invalid
+                 spwn/agents/neo/agent.yaml#deps
+                 → use skill/<name> (for spwn/skills/<name>.md), tool/<name> (for spwn/tools/<name>/), hook/<name> (for spwn/hooks/<name>.yaml), or command/<name> (for spwn/commands/<name>.md); spwn:<name> and github:<owner>/<repo> remain the two source-prefixed schemes
+
+            3 error(s), 0 warning(s), 0 info
+
