+++ tests/specs/cli/check/logs-without-world.spec.yaml
@@ -1,0 +1,10 @@
+description: world logs with no world id is refused by arity
+fixture: $FIXTURES/empty/
+env: isolated
+runs:
+    - command: world logs
+      exit: 1
+      stderr: |4+
+          
+            Error: accepts 1 arg(s), received 0
+
