+++ tests/specs/cli/check/mixed-tool-refs.spec.yaml
@@ -1,0 +1,10 @@
+description: catalog and local tool refs mixed in one manifest pass without error
+fixture: $FIXTURES/mixed-tool-refs/
+runs:
+    - command: check
+      exit: 0
+      stdout: |4+
+          
+            ✓  Project is valid
+               {{workdir}}/spwn.yaml
+
