+++ tests/specs/cli/check/outside-a-project.spec.yaml
@@ -1,0 +1,9 @@
+description: check outside a spwn project nudges the user at spwn init
+fixture: $FIXTURES/empty/
+runs:
+    - command: check
+      exit: 1
+      stderr: |4+
+          
+            Error: no spwn.yaml found — run `spwn init` first
+
