+++ tests/specs/cli/check/promoted-bare-ref.spec.yaml
@@ -1,0 +1,40 @@
+description: a bare name accepted on the cli lands as spwn:<name> in agent.yaml
+fixture: $FIXTURES/empty/
+runs:
+    - command: init
+      exit: 0
+      stdout: |4
+            Committed:
+              spwn.yaml                      # project manifest (worlds live inline)
+              spwn.lock                      # committed dependency pins (like package-lock.json)
+              spwn/agents/neo/               # starter agent
+              spwn/skills/focus.md           # example local skill  (skill/focus)
+              spwn/tools/greet/              # example local tool   (tool/greet)
+              spwn/hooks/session-banner.yaml # example local hook   (hook/session-banner)
+              spwn/commands/refactor.md      # example slash command (command/refactor)
+
+            Gitignored:
+              .spwn/                         # local state
+
+            Next:
+              spwn auth                      # verify your credentials
+              spwn agent neo                 # open an interactive session with neo
+      stderr: |4+
+          
+            ✓ Initialised spwn project {{string}}
+
+
+    - command: install python
+      exit: 0
+      stdout: |4
+            ✓  installed spwn:python
+               1 agent updated, spwn.lock pinned
+    - command: check
+      exit: 0
+      stdout: |4+
+          
+            ✓  Project is valid
+               {{workdir}}/spwn.yaml
+
+      files:
+          spwn/agents/neo/agent.yaml: { contains: 'spwn:python' }
