+++ tests/specs/cli/check/remove-without-name.spec.yaml
@@ -1,0 +1,10 @@
+description: agent rm with no name is refused by arity
+fixture: $FIXTURES/empty/
+env: isolated
+runs:
+    - command: agent rm
+      exit: 1
+      stderr: |4+
+          
+            Error: accepts 1 arg(s), received 0
+
