+++ tests/specs/cli/check/too-many-agent-names.spec.yaml
@@ -1,0 +1,10 @@
+description: agent create with three positional args is refused by arity
+fixture: $FIXTURES/empty/
+env: isolated
+runs:
+    - command: agent create a b c
+      exit: 1
+      stderr: |4+
+          
+            Error: accepts at most 1 arg(s), received 3
+
