+++ tests/specs/cli/check/unknown-builtin-tool.spec.yaml
@@ -1,0 +1,16 @@
+description: an agent referencing a non-existent built-in tool is flagged
+fixture: $FIXTURES/check-invalid-tool/
+runs:
+    - command: check
+      exit: 1
+      stdout: |4+
+          
+            {{workdir}}/spwn.yaml
+
+            errors
+              •  dependency "spwn:nonexistent" does not exist
+                 spwn/agents/neo/agent.yaml#deps
+                 → available built-ins: spwn:architect, spwn:build, spwn:cli, spwn:docker-cli, spwn:gh, spwn:git, spwn:mcp2cli, spwn:mempalace, spwn:node, spwn:notion, spwn:python, spwn:qmd, spwn:unix, spwn:x, spwn:claude-code, spwn:codex
+
+            1 error(s), 0 warning(s), 0 info
+
