+++ tests/specs/cli/check/unnamed-agent.spec.yaml
@@ -1,0 +1,14 @@
+description: agent create with no name picks a random planet name
+fixture: $FIXTURES/empty/
+env: isolated
+runs:
+    - command: agent create
+      exit: 0
+      stderr: |4+
+          
+            → Creating agent "{{string#agent}}"...
+            ✓ Created agent            {{string#agent}}
+            ✓ Created soul             SOUL.md
+
+            ✓ Spawn with: spwn up --agent {{string#agent}}
+
