+++ tests/specs/cli/check/usage-page.spec.yaml
@@ -1,0 +1,21 @@
+description: check --help prints the command usage
+fixture: $FIXTURES/empty/
+runs:
+    - command: check --help
+      exit: 0
+      stdout: |+
+          
+          Walks up from the current directory looking for spwn.yaml, then runs
+          every validation rule against the project. Reports issues grouped by
+          severity. Exits non-zero when errors are found (or warnings, with
+          --strict).
+
+          Usage:
+            spwn check [flags]
+
+          Flags:
+                --deep     Also run a compile dry-run and report renderer errors
+            -h, --help     help for check
+                --json     Emit results as structured JSON on stdout
+                --strict   Exit non-zero on warnings, not just errors
+
